small update deploy.bat

+++ sound_driver/deploy.bat
@@ -3,13 +3,14 @@
 :: Requiered dependency: pyinstaller
 :: See: https://pyinstaller.org/en/stable/
 
-echo Removing old deployment if exists...
-rmdir /s /q dist
+echo Removing old deployment (if exists)...
+if exist dist rmdir /s /q dist
 
 echo Installing sound_driver...
-pyinstaller sound_driver.py
-copy config.yaml dist\sound_driver\
+pyinstaller sound_driver.py > NUL 2>&1
+copy config.yaml dist\sound_driver\ > NUL 2>&1
 
-powershell Compress-Archive -Path dist\sound_driver -DestinationPath sound_driver.zip
+echo Creating a zip file...
+powershell Compress-Archive -Path dist\sound_driver -DestinationPath sound_driver.zip -Update
 
 echo Installation done
